+++ modules/engines/engine-hdb/pom.xml
@@ -21,6 +21,20 @@
       <groupId>com.sap.xsk</groupId>
       <artifactId>xsk-modules-parsers-all</artifactId>
       <version>0.5.0-SNAPSHOT</version>
+      <exclusions>
+        <exclusion>
+          <groupId>log4j</groupId>
+          <artifactId>log4j</artifactId>
+        </exclusion>
+        <exclusion>
+          <groupId>org.slf4j</groupId>
+          <artifactId>slf4j-api</artifactId>
+        </exclusion>
+        <exclusion>
+          <groupId>org.slf4j</groupId>
+          <artifactId>jcl-over-slf4j</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
     <dependency>
       <groupId>com.sap.xsk</groupId>
@@ -32,6 +46,12 @@
       <groupId>org.eclipse.dirigible</groupId>
       <artifactId>dirigible-engine-javascript-graalvm</artifactId>
       <version>${dirigible.version}</version>
+      <exclusions>
+        <exclusion>
+          <groupId>org.slf4j</groupId>
+          <artifactId>jcl-over-slf4j</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
     <dependency>
       <groupId>org.eclipse.dirigible</groupId>
@@ -54,26 +74,56 @@
       <groupId>org.eclipse.xtext</groupId>
       <artifactId>org.eclipse.xtext.testing</artifactId>
       <version>${xtextVersion}</version>
+      <exclusions>
+        <exclusion>
+          <groupId>log4j</groupId>
+          <artifactId>log4j</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
     <dependency>
       <groupId>org.eclipse.xtext</groupId>
       <artifactId>org.eclipse.xtext.xbase.testing</artifactId>
       <version>${xtextVersion}</version>
+      <exclusions>
+        <exclusion>
+          <groupId>log4j</groupId>
+          <artifactId>log4j</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
     <dependency>
       <groupId>org.eclipse.xtext</groupId>
       <artifactId>org.eclipse.xtext</artifactId>
       <version>${xtextVersion}</version>
+      <exclusions>
+        <exclusion>
+          <groupId>log4j</groupId>
+          <artifactId>log4j</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
     <dependency>
       <groupId>org.eclipse.xtext</groupId>
       <artifactId>org.eclipse.xtext.xbase</artifactId>
       <version>${xtextVersion}</version>
+      <exclusions>
+        <exclusion>
+          <groupId>log4j</groupId>
+          <artifactId>log4j</artifactId>
+        </exclusion>
+      </exclusions>
     </dependency>
     <dependency>
       <groupId>org.eclipse.xtext</groupId>
       <artifactId>org.eclipse.xtext.xtext.generator</artifactId>
       <version>${xtextVersion}</version>
+      <exclusions>
+        <exclusion>
+          <groupId>log4j</groupId>
+          <artifactId>log4j</artifactId>
+        </exclusion>
+      </exclusions>
       <optional>true</optional>
     </dependency>
 
@@ -85,6 +135,14 @@
         <exclusion>
           <groupId>org.eclipse.dirigible</groupId>
           <artifactId>dirigible-database-dynamic</artifactId>
+        </exclusion>
+        <exclusion>
+          <groupId>org.slf4j</groupId>
+          <artifactId>slf4j-log4j12</artifactId>
+        </exclusion>
+        <exclusion>
+          <groupId>org.slf4j</groupId>
+          <artifactId>jul-to-slf4j</artifactId>
         </exclusion>
       </exclusions>
     </dependency>
@@ -103,16 +161,16 @@
       <artifactId>dirigible-repository-api</artifactId>
       <version>${dirigible.version}</version>
     </dependency>
-    <dependency>
-      <groupId>org.eclipse.dirigible</groupId>
-      <artifactId>dirigible-bpm</artifactId>
-      <version>${dirigible.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.dirigible</groupId>
-      <artifactId>dirigible-ext-acorn</artifactId>
-      <version>${dirigible.version}</version>
-    </dependency>
+    <!-- <dependency>
+       <groupId>org.eclipse.dirigible</groupId>
+       <artifactId>dirigible-bpm</artifactId>
+       <version>${dirigible.version}</version>
+     </dependency>
+      <dependency>
+        <groupId>org.eclipse.dirigible</groupId>
+        <artifactId>dirigible-ext-acorn</artifactId>
+        <version>${dirigible.version}</version>
+      </dependency>-->
 
     <dependency>
       <groupId>javax.servlet</groupId>
@@ -208,6 +266,7 @@
       <groupId>com.sap.cloud.db.jdbc</groupId>
       <artifactId>ngdbc</artifactId>
       <version>2.8.12</version>
+      <scope>test</scope>
     </dependency>
 
     <dependency>
@@ -275,6 +334,40 @@
                   <goal>verify</goal>
                 </goals>
                 <inherited>false</inherited>
+              </execution>
+            </executions>
+          </plugin>
+          <plugin>
+            <groupId>org.codehaus.mojo</groupId>
+            <artifactId>build-helper-maven-plugin</artifactId>
+            <version>1.9.1</version>
+            <executions>
+              <execution>
+                <id>add-integration-test-sources</id>
+                <phase>generate-test-sources</phase>
+                <goals>
+                  <goal>add-test-source</goal>
+                </goals>
+                <configuration>
+                  <sources>
+                    <source>src/itest/java</source>
+                  </sources>
+                </configuration>
+              </execution>
+              <execution>
+                <id>add-integration-test-resources</id>
+                <phase>generate-test-resources</phase>
+                <goals>
+                  <goal>add-test-resource</goal>
+                </goals>
+                <configuration>
+                  <resources>
+                    <resource>
+                      <filtering>true</filtering>
+                      <directory>src/itest/resources</directory>
+                    </resource>
+                  </resources>
+                </configuration>
               </execution>
             </executions>
           </plugin>
