+++ modules/engines/engine-hdb/src/itest/java/com.sap.xsk.hdb.ds.itest/hdbtable/XSKHDBTableParserMySQLITTest.java
@@ -9,7 +9,7 @@
  * SPDX-FileCopyrightText: 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sap.xsk.hdb.ds.test.itest.hdbview;
+package com.sap.xsk.hdb.ds.itest.hdbtable;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
@@ -17,14 +17,13 @@ import com.google.inject.Key;
 import com.google.inject.name.Names;
 import com.sap.xsk.hdb.ds.api.XSKDataStructuresException;
 import com.sap.xsk.hdb.ds.facade.IXSKHDBCoreFacade;
+import com.sap.xsk.hdb.ds.itest.module.XSKHDBTestModule;
 import com.sap.xsk.hdb.ds.test.itest.model.JDBCModel;
-import com.sap.xsk.hdb.ds.test.itest.module.XSKHDBTestModule;
-import com.sap.xsk.utils.XSKHDBUtils;
 import org.eclipse.dirigible.core.scheduler.api.SynchronizationException;
 import org.eclipse.dirigible.repository.local.LocalResource;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.containers.MySQLContainer;
 import javax.sql.DataSource;
 import java.io.IOException;
 import java.sql.Connection;
@@ -32,12 +31,11 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-public class XSKHDBViewParserPostgreSQLITTest {
+public class XSKHDBTableParserMySQLITTest {
 
-  private static PostgreSQLContainer jdbcContainer;
+  private static MySQLContainer jdbcContainer;
   private static Connection connection;
   private static IXSKHDBCoreFacade facade;
 
@@ -45,7 +43,7 @@ public class XSKHDBViewParserPostgreSQLITTest {
   @BeforeClass
   public static void setUp() throws SQLException {
     jdbcContainer =
-        new PostgreSQLContainer<>("postgres:alpine");
+        new MySQLContainer<>("mysql:5.7");
     jdbcContainer.start();
     JDBCModel model = new JDBCModel(jdbcContainer.getDriverClassName(), jdbcContainer.getJdbcUrl(), jdbcContainer.getUsername(),
         jdbcContainer.getPassword());
@@ -54,26 +52,19 @@ public class XSKHDBViewParserPostgreSQLITTest {
     facade = injector.getInstance(Key.get(IXSKHDBCoreFacade.class, Names.named("xskHDBCoreFacade")));
   }
 
-
   @Test
-  public void testHDBViewCreate() throws XSKDataStructuresException, SynchronizationException, IOException, SQLException {
-    Statement stmt = connection.createStatement();
-    stmt.executeUpdate("create table \"public\".\"acme.com.test.tables::MY_TABLE1\"(Column1 integer,Column2 integer)");
-    stmt.executeUpdate("create table \"public\".\"acme.com.test.views::MY_VIEW1\"(Column1 integer,Column2 integer)");
+  public void testHDBTableCreate() throws XSKDataStructuresException, SynchronizationException, IOException, SQLException {
     LocalResource resource = XSKHDBTestModule.getResources("/usr/local/target/dirigible/repository/root",
-        "/registry/public/hdbview-itest/SamplePostgreXSClassicView.hdbview",
-        "/hdbview-itest/SamplePostgreXSClassicView.hdbview");
+        "/registry/public/hdbtable-itest/SamplePostgreXSClassicTable.hdbtable",
+            "/hdbtable-itest/SamplePostgreXSClassicTable.hdbtable");
 
     this.facade.handleResourceSynchronization(resource);
     this.facade.updateEntities();
 
-    ResultSet rs = stmt
-        .executeQuery(String.format("SELECT COUNT(*) as rawsCount FROM \"%s\"", "hdbview-itest::SamplePostgreXSClassicView"));
+    Statement stmt = connection.createStatement();
+    ResultSet rs = stmt.executeQuery("SELECT COUNT(*) as rawsCount FROM `hdbtable-itest::SamplePostgreXSClassicTable`");
     assertTrue(rs.next());
-    assertEquals(0, rs.getInt("rawsCount"));
-    stmt.executeUpdate(
-        String.format("DROP VIEW %s", XSKHDBUtils.escapeArtifactName(connection, "hdbview-itest::SamplePostgreXSClassicView")));
-    stmt.executeUpdate("drop table \"public\".\"acme.com.test.tables::MY_TABLE1\"");
-    stmt.executeUpdate("drop table \"public\".\"acme.com.test.views::MY_VIEW1\"");
+    assertTrue(0 == rs.getInt("rawsCount"));
+    stmt.executeUpdate("DROP TABLE `hdbtable-itest::SamplePostgreXSClassicTable`");
   }
 }
