+++ modules/engines/engine-hdb/src/itest/java/com.sap.xsk.hdb.ds.itest/hdbview/XSKHDBViewParserHanaITTest.java
@@ -1,4 +1,4 @@
-/*
+package com.sap.xsk.hdb.ds.itest.hdbview;/*
  * Copyright (c) 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
  *
  * All rights reserved. This program and the accompanying materials
@@ -9,16 +9,27 @@
  * SPDX-FileCopyrightText: 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sap.xsk.hdb.ds.test.itest.hdbview;
-
+///*
+// * Copyright (c) 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
+// *
+// * All rights reserved. This program and the accompanying materials
+// * are made available under the terms of the Apache License, v2.0
+// * which accompanies this distribution, and is available at
+// * http://www.apache.org/licenses/LICENSE-2.0
+// *
+// * SPDX-FileCopyrightText: 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
+// * SPDX-License-Identifier: Apache-2.0
+// */
+//package com.sap.xsk.hdb.ds.itest.hdbview;
+//
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.Key;
 import com.google.inject.name.Names;
 import com.sap.xsk.hdb.ds.api.XSKDataStructuresException;
 import com.sap.xsk.hdb.ds.facade.IXSKHDBCoreFacade;
+import com.sap.xsk.hdb.ds.itest.module.XSKHDBTestModule;
 import com.sap.xsk.hdb.ds.test.itest.model.JDBCModel;
-import com.sap.xsk.hdb.ds.test.itest.module.XSKHDBTestModule;
 import com.sap.xsk.utils.XSKHDBUtils;
 import org.eclipse.dirigible.commons.config.Configuration;
 import org.eclipse.dirigible.core.scheduler.api.SynchronizationException;
@@ -33,7 +44,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-import static com.sap.xsk.hdb.ds.test.itest.utils.TestConstants.*;
+import static com.sap.xsk.hdb.ds.itest.utils.TestConstants.*;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
