+++ modules/engines/engine-hdb/src/itest/java/com.sap.xsk.hdb.ds.itest/hdbview/XSKHDBViewParserMySQLITTest.java
@@ -9,7 +9,7 @@
  * SPDX-FileCopyrightText: 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sap.xsk.hdb.ds.test.itest.hdbtable;
+package com.sap.xsk.hdb.ds.itest.hdbview;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
@@ -18,13 +18,12 @@ import com.google.inject.name.Names;
 import com.sap.xsk.hdb.ds.api.XSKDataStructuresException;
 import com.sap.xsk.hdb.ds.facade.IXSKHDBCoreFacade;
 import com.sap.xsk.hdb.ds.test.itest.model.JDBCModel;
-import com.sap.xsk.hdb.ds.test.itest.module.XSKHDBTestModule;
+import com.sap.xsk.hdb.ds.itest.module.XSKHDBTestModule;
 import org.eclipse.dirigible.core.scheduler.api.SynchronizationException;
 import org.eclipse.dirigible.repository.local.LocalResource;
-import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.containers.MySQLContainer;
 import javax.sql.DataSource;
 import java.io.IOException;
 import java.sql.Connection;
@@ -32,11 +31,12 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
-public class XSKHDBTableParserPostgreSQLITTest {
+public class XSKHDBViewParserMySQLITTest {
 
-  private static PostgreSQLContainer jdbcContainer;
+  private static MySQLContainer jdbcContainer;
   private static Connection connection;
   private static IXSKHDBCoreFacade facade;
 
@@ -44,7 +44,7 @@ public class XSKHDBTableParserPostgreSQLITTest {
   @BeforeClass
   public static void setUp() throws SQLException {
     jdbcContainer =
-        new PostgreSQLContainer<>("postgres:alpine");
+        new MySQLContainer<>("mysql:5.7");
     jdbcContainer.start();
     JDBCModel model = new JDBCModel(jdbcContainer.getDriverClassName(), jdbcContainer.getJdbcUrl(), jdbcContainer.getUsername(),
         jdbcContainer.getPassword());
@@ -53,25 +53,24 @@ public class XSKHDBTableParserPostgreSQLITTest {
     facade = injector.getInstance(Key.get(IXSKHDBCoreFacade.class, Names.named("xskHDBCoreFacade")));
   }
 
-  @AfterClass
-  public static void cleanUp() {
-    jdbcContainer.stop();
-  }
-
   @Test
-  public void testHDBTableCreate() throws XSKDataStructuresException, SynchronizationException, IOException, SQLException {
+  public void testHDBViewCreate() throws XSKDataStructuresException, SynchronizationException, IOException, SQLException {
+    Statement stmt = connection.createStatement();
+    stmt.executeUpdate("create table `test`.`acme.com.test.tables::MY_TABLE1`(Column1 integer,Column2 integer)");
+    stmt.executeUpdate("create table `test`.`acme.com.test.views::MY_VIEW1`(Column1 integer,Column2 integer)");
+
     LocalResource resource = XSKHDBTestModule.getResources("/usr/local/target/dirigible/repository/root",
-        "/registry/public/hdbtable-itest/SamplePostgreXSClassicTable.hdbtable",
-        "/hdbtable-itest/SamplePostgreXSClassicTable.hdbtable");
+        "/registry/public/hdbview-itest/SampleMySQLXSClassicView.hdbview",
+            "/hdbview-itest/SampleMySQLXSClassicView.hdbview");
 
     this.facade.handleResourceSynchronization(resource);
     this.facade.updateEntities();
 
-    Statement stmt = connection.createStatement();
-    ResultSet rs = stmt
-        .executeQuery(String.format("SELECT COUNT(*) as rawsCount FROM \"%s\"", "hdbtable-itest::SamplePostgreXSClassicTable"));
+    ResultSet rs = stmt.executeQuery(String.format("SELECT COUNT(*) as rawsCount FROM %s", "`hdbview-itest::SampleMySQLXSClassicView`"));
     assertTrue(rs.next());
     assertEquals(0, rs.getInt("rawsCount"));
-    stmt.executeUpdate(String.format("DROP TABLE \"%s\"", "hdbtable-itest::SamplePostgreXSClassicTable"));
+    stmt.executeUpdate(String.format("DROP VIEW %s", "`hdbview-itest::SampleMySQLXSClassicView`"));
+    stmt.executeUpdate("drop table `test`.`acme.com.test.tables::MY_TABLE1`");
+    stmt.executeUpdate("drop table `test`.`acme.com.test.views::MY_VIEW1`");
   }
 }
