+++ modules/engines/engine-hdb/src/itest/java/com.sap.xsk.hdb.ds.itest/hdbview/XSKHDBViewParserPostgreSQLITTest.java
@@ -9,7 +9,7 @@
  * SPDX-FileCopyrightText: 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sap.xsk.hdb.ds.test.itest.hdbsequence;
+package com.sap.xsk.hdb.ds.itest.hdbview;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
@@ -17,14 +17,13 @@ import com.google.inject.Key;
 import com.google.inject.name.Names;
 import com.sap.xsk.hdb.ds.api.XSKDataStructuresException;
 import com.sap.xsk.hdb.ds.facade.IXSKHDBCoreFacade;
+import com.sap.xsk.hdb.ds.itest.module.XSKHDBTestModule;
 import com.sap.xsk.hdb.ds.test.itest.model.JDBCModel;
-import com.sap.xsk.hdb.ds.test.itest.module.XSKHDBTestModule;
+import com.sap.xsk.utils.XSKHDBUtils;
 import org.eclipse.dirigible.core.scheduler.api.SynchronizationException;
 import org.eclipse.dirigible.repository.local.LocalResource;
-import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.testcontainers.containers.Network;
 import org.testcontainers.containers.PostgreSQLContainer;
 import javax.sql.DataSource;
 import java.io.IOException;
@@ -32,18 +31,11 @@ import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.List;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
-/**
- * XSKHDBSequenceParserITCase will test only sequence
- * creation due existing sequence drop logic that is never invoked.
- * Existing sequence alter logic is not checked due db limitations on retrieving sequence info.
- */
-public class XSKHDBSequenceParserPostgreSQLITTest {
+public class XSKHDBViewParserPostgreSQLITTest {
 
   private static PostgreSQLContainer jdbcContainer;
   private static Connection connection;
@@ -52,7 +44,6 @@ public class XSKHDBSequenceParserPostgreSQLITTest {
 
   @BeforeClass
   public static void setUp() throws SQLException {
-    Network network = Network.newNetwork();
     jdbcContainer =
         new PostgreSQLContainer<>("postgres:alpine");
     jdbcContainer.start();
@@ -63,33 +54,26 @@ public class XSKHDBSequenceParserPostgreSQLITTest {
     facade = injector.getInstance(Key.get(IXSKHDBCoreFacade.class, Names.named("xskHDBCoreFacade")));
   }
 
-  @AfterClass
-  public static void cleanUp() {
-    jdbcContainer.stop();
-  }
 
   @Test
-  public void testHDBSequenceCreate() throws XSKDataStructuresException, SynchronizationException, IOException, SQLException {
+  public void testHDBViewCreate() throws XSKDataStructuresException, SynchronizationException, IOException, SQLException {
+    Statement stmt = connection.createStatement();
+    stmt.executeUpdate("create table \"public\".\"acme.com.test.tables::MY_TABLE1\"(Column1 integer,Column2 integer)");
+    stmt.executeUpdate("create table \"public\".\"acme.com.test.views::MY_VIEW1\"(Column1 integer,Column2 integer)");
     LocalResource resource = XSKHDBTestModule.getResources("/usr/local/target/dirigible/repository/root",
-        "/registry/public/sequence-itest/SampleSequence_HanaXSClassic.hdbsequence",
-        "/sequence-itest/SampleSequence_HanaXSClassic.hdbsequence");
+        "/registry/public/hdbview-itest/SamplePostgreXSClassicView.hdbview",
+        "/hdbview-itest/SamplePostgreXSClassicView.hdbview");
 
     this.facade.handleResourceSynchronization(resource);
     this.facade.updateEntities();
 
-    Statement stmt = connection.createStatement();
-    List<String> dbSequences = new ArrayList<>();
-    ResultSet rs = stmt.executeQuery("SELECT  relname sequence_name FROM  pg_class WHERE  relkind = 'S'");
-    while (rs.next()) {
-      dbSequences.add(rs.getString("sequence_name"));
-    }
-    assertEquals(1, dbSequences.size());
-    assertEquals("sequence-itest::SampleSequence_HanaXSClassic", dbSequences.get(0));
-
-    stmt.executeUpdate(String.format("DROP SEQUENCE \"%s\"", dbSequences.get(0)));
-    rs = stmt.executeQuery("SELECT  relname sequence_name FROM  pg_class WHERE  relkind = 'S'");
-    assertFalse(rs.next());
+    ResultSet rs = stmt
+        .executeQuery(String.format("SELECT COUNT(*) as rawsCount FROM \"%s\"", "hdbview-itest::SamplePostgreXSClassicView"));
+    assertTrue(rs.next());
+    assertEquals(0, rs.getInt("rawsCount"));
+    stmt.executeUpdate(
+        String.format("DROP VIEW %s", XSKHDBUtils.escapeArtifactName(connection, "hdbview-itest::SamplePostgreXSClassicView")));
+    stmt.executeUpdate("drop table \"public\".\"acme.com.test.tables::MY_TABLE1\"");
+    stmt.executeUpdate("drop table \"public\".\"acme.com.test.views::MY_VIEW1\"");
   }
-
-
 }
