+++ modules/engines/engine-hdb/src/itest/java/com.sap.xsk.hdb.ds.itest/module/XSKHDBTestModule.java
@@ -9,12 +9,12 @@
  * SPDX-FileCopyrightText: 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sap.xsk.hdb.ds.test.itest.module;
+package com.sap.xsk.hdb.ds.itest.module;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
+import com.sap.xsk.hdb.ds.itest.hdbsequence.XSKHDBSequenceParserPostgreSQLITTest;
 import com.sap.xsk.hdb.ds.module.XSKHDBModule;
-import com.sap.xsk.hdb.ds.test.itest.hdbsequence.XSKHDBSequenceParserPostgreSQLITTest;
 import com.sap.xsk.hdb.ds.test.itest.model.JDBCModel;
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.eclipse.dirigible.repository.api.RepositoryPath;
