+++ modules/engines/engine-hdb/src/itest/java/com.sap.xsk.hdb.ds.itest/utils/TestConstants.java
@@ -9,7 +9,7 @@
  * SPDX-FileCopyrightText: 2019-2021 SAP SE or an SAP affiliate company and XSK contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sap.xsk.hdb.ds.test.itest.utils;
+package com.sap.xsk.hdb.ds.itest.utils;
 
 import org.eclipse.dirigible.commons.config.Configuration;
