+++ modules/engines/engine-hdb/src/itest/resources/hdbtable-itest/SamplePostgreXSClassicTable.hdbtable
