+++ modules/engines/engine-hdb/src/itest/resources/hdbview-itest/SampleMySQLXSClassicView.hdbview
