+++ modules/engines/engine-hdb/src/itest/resources/sequence-itest/SampleSequence_HanaXSClassic.hdbsequence
